Vendor the Bootstrap CSS.
Rather than relying on an external path with questionable longevity (the
Bootstrap homepage hosted on GitHub Pages), include the Bootstrap CSS
under the `vendor` directory and require it in the `application.css`.

This also solves the issue of serving up the application layout on SSL
hosts in browsers like Chrome that prohibit the loading of insecure CSS
resources. Since GitHub Pages doesn't provide an SSL host, the layout
can't simply refer to a protocol-relative URL.