Raise error if configuration is not set
In case configuration for Doorkeeper was never set then trying to
reach the configuration would result in an error. That way the
users would easily see that they need to add doorkeeper initializer
or change it in order for it to work instead of getting
'undefined_method for nil' errors.