Extract ORMs from doorkeeper
This PR removes doorkeeper ORM specifics, in favor of extending it with
other rubygems.

The specs defined in this project are still considered canonical, and
the goal is for extension rubygems to set doorkeeper as a git submodule,
and running its test suite against the specific ORM that is being
extended.

See https://github.com/doorkeeper-gem/doorkeeper-orms/pull/2