Update doorkeeper to be used with grape >= 0.10 (#915)
Grape has since long discouraged the use of options[:route_options] for custom additions. Since version 0.10 it cannot be used anymore like setup in the helpers in doorkeeper (grape is currently at 0.19.1).

This PR fixes the behaviour to properly work with grape, and describes a bit better in the example how you could actually use it and set it up (if you want I can elaborate more in the readme).

I think with this example and the change, most people should have enough with doorkeeper as is, and not need wine_bouncer at all.